requirements.txt
fastapi
uvicorn
python-dotenv
openai
langchain
langchain-community
langchain-text-splitters
langchain-openai
faiss-cpu
numpy
pypdf